+++ src/main/java/com/hardwareInfo/hardwareInfo/services/SubVendorService.java
@@ -1,45 +1,45 @@
 package com.hardwareInfo.hardwareInfo.services;
 
 import com.hardwareInfo.hardwareInfo.entities.SubVendorEntity;
-import com.hardwareInfo.hardwareInfo.entities.VendorEntity;
+import com.hardwareInfo.hardwareInfo.exceptions.SubVendorNotFoundException;
 import com.hardwareInfo.hardwareInfo.repositories.SubVendorRepository;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 
 import java.util.List;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
 public class SubVendorService {
     private final SubVendorRepository subVendorRepository;
+    private final ModelMapper modelMapper;
 
-    public SubVendorEntity createVendor(SubVendorEntity subVendorEntity){
-
-        return subVendorRepository.save(subVendorEntity);
-    }
-    public Optional<SubVendorEntity> getVendorByID(Long id){
-        return subVendorRepository.findById(id);
-    }
-
-    public List<SubVendorEntity> getAllVendors(){
+    public List<SubVendorEntity> getAllSubVendors() {
         return subVendorRepository.findAll();
     }
-    @PutMapping
-    public Optional<SubVendorEntity> updateVendor(Long id, SubVendorEntity newVendor){
 
-        return subVendorRepository.findById(id).map( subVendor-> {
-            subVendor.setGpuVendor(newVendor.getGpuVendor());
-            subVendor.setGraphicsCards(newVendor.getGraphicsCards());
-            return subVendor;
-        });
+    public SubVendorEntity getSubVendorById(Long id) {
+        return subVendorRepository.findById(id)
+                .orElseThrow(() -> new SubVendorNotFoundException("Sub Vendor not found with id + " + id));
     }
-    @DeleteMapping
-    public void deleteVendor(Long id){
 
+    public SubVendorEntity createSubVendor(SubVendorEntity subVendorEntity) {
+        return subVendorRepository.save(subVendorEntity);
+    }
+
+    public SubVendorEntity updateSubVendor(Long id, SubVendorEntity newSubVendorEntity) {
+        SubVendorEntity existingSubVendorEntity = subVendorRepository.findById(id)
+                .orElseThrow(() -> new SubVendorNotFoundException("Sub Vendor not found with id + " + id));
+
+        modelMapper.map(newSubVendorEntity, existingSubVendorEntity);
+        return subVendorRepository.save(existingSubVendorEntity);
+    }
+
+    public void deleteSubVendor(Long id) {
+        if(!subVendorRepository.existsById(id)) {
+            throw new SubVendorNotFoundException("Sub Vendor not found with id + " + id);
+        }
         subVendorRepository.deleteById(id);
     }
-
 }
