+++ src/main/java/com/hardwareInfo/hardwareInfo/services/VramTypeService.java
@@ -1,46 +1,45 @@
 package com.hardwareInfo.hardwareInfo.services;
 
 import com.hardwareInfo.hardwareInfo.entities.VramTypeEntity;
+import com.hardwareInfo.hardwareInfo.exceptions.VramTypeNotFoundException;
 import com.hardwareInfo.hardwareInfo.repositories.VramTypeRepository;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 
 import java.util.List;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
 public class VramTypeService {
     private final VramTypeRepository vramTypeRepository;
+    private final ModelMapper modelMapper;
+
+    public List<VramTypeEntity> getAllVramTypes() {
+        return vramTypeRepository.findAll();
+    }
+
+    public VramTypeEntity getVramTypeById(Long id) {
+        return vramTypeRepository.findById(id)
+                .orElseThrow(() -> new VramTypeNotFoundException("Vram Type not found with id + " + id));
+    }
 
     public VramTypeEntity createVramType(VramTypeEntity vramTypeEntity) {
-
         return vramTypeRepository.save(vramTypeEntity);
     }
-    public List<VramTypeEntity> getAllVramTypes(){
 
-        return  vramTypeRepository.findAll();
+    public VramTypeEntity updateVramType(Long id, VramTypeEntity newVramTypeEntity) {
+        VramTypeEntity existingVramTypeEntity = vramTypeRepository.findById(id)
+                .orElseThrow(() -> new VramTypeNotFoundException("Vram Type not found with id + " + id));
+
+        modelMapper.map(newVramTypeEntity, existingVramTypeEntity);
+        return vramTypeRepository.save(existingVramTypeEntity);
     }
 
-    public Optional<VramTypeEntity> getVramTypeByID(Long id){
-
-        return vramTypeRepository.findById(id);
+    public void deleteVramType(Long id) {
+        if(!vramTypeRepository.existsById(id)) {
+            throw new VramTypeNotFoundException("Vram Type not found with id + " + id);
+        }
+        vramTypeRepository.deleteById(id);
     }
-    @PutMapping
-    public Optional<VramTypeEntity> updateVramType(Long id, VramTypeEntity newVramType){
-        return vramTypeRepository.findById(id).map( vramType -> {
-            vramType.setVramType(newVramType.getVramType());
-            return vramTypeRepository.save(vramType);
-        });
-
-    }
-
-
-    @DeleteMapping
-    public void deleteVramType(Long id){
-         vramTypeRepository.deleteById(id);
-    }
-
 }
